+++ app/src/main/java/com/yogeshpaliyal/comrade/reciever/handlers/BackupBroadcastHandler.kt
@@ -4,17 +4,23 @@ import android.content.Context
 import android.content.Intent
 import android.content.pm.PackageManager
 import android.net.Uri
+import com.yogeshpaliyal.comrade.common.IA_BACKUP_ADDED_TO_QUEUE
 import com.yogeshpaliyal.comrade.common.IA_BACKUP_REQUEST
 import com.yogeshpaliyal.comrade.common.SHARING_CONTENT_URI
 import com.yogeshpaliyal.comrade.types.BackupStatus
+import data.ComradeBackupQueries
 import java.io.File
 import java.io.IOException
+import javax.inject.Inject
 
-class BackupBroadcastHandler: IBroadcastReceiverHandler {
+class BackupBroadcastHandler @Inject constructor(): IBroadcastReceiverHandler {
     override val type: String
         get() = IA_BACKUP_REQUEST
 
-    override fun handleAction(context: Context, intent: Intent) {
+    @Inject
+    lateinit var queries: ComradeBackupQueries
+
+    override fun handleAction(context: Context, intent: Intent, callingApp: String) {
         val contentUri = intent.getStringExtra(SHARING_CONTENT_URI)
 
         if (contentUri != null) {
@@ -47,5 +53,12 @@ class BackupBroadcastHandler: IBroadcastReceiverHandler {
                 e.printStackTrace()
             }
         }
+    }
+
+    private fun backupAddedToQueue(context: Context?, targetApp: String) {
+        val mIntent = Intent()
+        mIntent.action = IA_BACKUP_ADDED_TO_QUEUE
+        mIntent.`package` = targetApp
+        context?.sendBroadcast(mIntent)
     }
 }
