+++ app/src/main/java/com/yogeshpaliyal/comrade/reciever/handlers/BroadcastReceiverHandler.kt
@@ -5,5 +5,5 @@ import android.content.Intent
 
 interface IBroadcastReceiverHandler {
     val type: String
-    fun handleAction(context: Context, intent: Intent)
+    fun handleAction(context: Context, intent: Intent, callingApp: String)
 }
